Added ability to specify .h files on Chapel command line.  This causes the .h file to be #included by the generated Chapel C code, and is the preferred way to supply prototypes for .c and .o files specified on the command-line to the back-end C compiler. (Previously, we were generating our own prototypes which may or may not have matched the external C code).
This change makes it possible for Jim's madness code
to call sprintf directly rather than via a C wrapper,
but also generates warnings for the C compile step
(in both UTS and Madness) because the string type in
Chapel is a const char*, and the external routines
used by both benchmarks modify the string in place.
Possible fixes would be to supply a fixed-size
string(len) type, which we've conceived of as being
modifiable in-place (this would also clean up some
of his uses of strings), to provide an opaque type
in Chapel and have the C code create these opaque
types for the user, to use a _ddata buffer, or to
cast out the const-ness in the C wrapper code.

In spite of these warnings, checking in this change
because having the compiler complain about constness
is a less-serious offense than lying to the C
compiler about a routine's interface.  (Jim pointed
out that C++ compilers might do more than just complain
about casting away constness, but they probably would
not have allowed the link to happen either, as the
function names would've been munged differently for
the const and non-const cases.

In this change, I also permitted external functions to
be overloaded (so that multiple varieties of sprintf()
could be prototyped for example), and disabled munging
of external symbol names (since we can't munge the
external sources).  A good next step would be to permit
external functions to be generic and/or vararg functions,
both of which should be simpler now that we're not
creating the prototypes -- but that's another check-in.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11518 3a8e244f-b0f2-452b-bcba-4c88e055c3ca