Added the ability to do Makefile-driven testing within a subdirectory of the testing system.  Working in this mode drops most of the capabilities of the testing system for a directory, turning control over to the user for better or worse.  With additional work, more of the command-line flags used by the testing system could be fed into a testing Makefile using well-defined identifiers, but I haven't set up those hooks as of yet.  Please send feature requests and bug reports to me as you encounter them.
Used the Makefile-driven testing capability to test
Franz's fft code using a Makefile, since it had lots
of source files but only one conceptual test.  Removed
the .notest files which are no longer necessary.

Here's the documentation of this feature, from the
comments in start_test:

An alternate way of setting up a subdirectory for testing is to
use a Makefile-driven test.  This is most useful for tests that
require several source files (to avoid littering the directory
with .notest files), or a very complicated build procedure.  To
set up such a test, the following steps should be taken:

1) add a file USEMAKE to the directory

2) add a Makefile to the directory that contains a "test" target:
   - the Makefile should refer to the chpl compiler using $(CHPL),
     which will be fed to the Makefile by the testing script.
   - make the Makefile execute any binaries using -nl 1 (or
     something even more intelligent to make it work well with
     parallel testing.
   - you will probably want to squelch the Makefile's echoing of
     commands using the @ prefix
   - and to avoid breaking due to error status codes using the -
     prefix?

3) add a file make.good to the directory indicating what the output
   of "make test" ought to be

Note that for such directories, performance testing is still done
using the usual mechanism: by providing .perfkeys, .perfcompopts,
and .perfexecopts files.  This could be replaced by a "perftest"
target if there is a compelling reason to do so.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11527 3a8e244f-b0f2-452b-bcba-4c88e055c3ca