Before this check-in, one could prototype external functions whose formals were missing their types and/or external varargs functions, but the compiler would at some point ignore the fact that they were extern, decide that you actually wanted an internal function with an empty statement list and create that function for you.  Boo!  :(
With this check-in, such external functions can
be declared.  For example, you could declare an
external function that took four arguments of
unspecified type (because you were lazy?  because
you're going to compile against C++ and use an
overloaded function?) using:

_extern def foo(w, x, y, z);

Or, you could define a printf that took 4 arguments
using:

_extern def printf(fmt: string, w, x, y, z);

But, even better, you can define the varargs version
of print as:

_extern def printf(fmt: string, vals:...?numvals);

Added three tests of this behavior.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11534 3a8e244f-b0f2-452b-bcba-4c88e055c3ca