Submitting two feature request futures (and two workarounds showing how things currently would be done) for features requested by Jim over lunch which seemed reasonable to me.
The first regards variables being declared
within a cobegin statement escaping that scope
(i.e., put into the cobegin's parent's lexical
scope) to avoid forcing the variables to be
declared a priori, typed, be left const-less, etc.
This is something I've been frustrated by when
using cobegins as well.  The .future file discusses
the tradeoffs that I understand.

The second is regarding a var function's ability
to return const variables (or, more generally,
non-lvalues) when it is in getter mode rather
than setter mode.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11560 3a8e244f-b0f2-452b-bcba-4c88e055c3ca