Updated range and domain "slicing" to compute the intersection of the slicer range and the slicee range.  Range and domain slicing can no longer result in an out-of-bounds error.
This is not yet correctly implemented for the case where the resulting range is
unbounded in the upper bound or the case where the stride is negative.

Updated several test cases and removed a .future.

Fixed the spec section on domain slicing but this needs updating.
Also removed a redunandant "Section" word as in "In Section
(SectionSymbol)...".  Also note that the implementation of domain
slicing needs help.  This should be in the arithmetic domain, not the
domain wrapper, and should allow for slicing by domains, not just
ranges.  Is there a reason why not?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11648 3a8e244f-b0f2-452b-bcba-4c88e055c3ca