Removed uses of strtoll and strtoull from the runtime due to their general non-portability on many platforms (sadly -- they are clean routines).
Did this by rewriting string_to_int* and string_to_uint*
routines to call strtol for *=8, 16, 32, and to call
sscanf for *=64.

Improved the robustness of these routines by having
them complain bitterly (maybe too bitterly) if the
string has extra stuff tacked onto the end.  This
should help with some other futures we have
regarding sloppiness in the runtime in doing
string-to-int conversions.  If the invokee of a
conversion wants to generate their own error message,
they can call the *_precise() version of the call
which takes a second argument indicating whether or
not the conversion was valid rather than complaining.

In the 64-bit cases, unfortunately, the validity of
the string is poor, as I haven't determined a way
yet to determine whether or not sscanf() consumed
the whole string (could've sworn that there used
to be some way to do this, but maybe it got deprecated).

This change also makes the following improvements:

- runtime string-to-int conversions used to use
  strtoull rather than strtoll, which implies that
  the sign probably would've gotten dropped;  I
  hypothesize that we never tested this.

- consolidates all calls to strto* into the chplcast
  code so that assumptions about integer sizes are
  not spread throughout the runtime.

It does not make the following improvement:

- we still have big assumptions about how big
  integers are in the runtime which should be
  generalized or parameterized somehow.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11687 3a8e244f-b0f2-452b-bcba-4c88e055c3ca