Improved string-to-int conversion routines to be pickier about extra characters at the ends of strings, and to generate better error messages about what caused the conversion to fail.  Added a bunch of tests to lock in the behavior I just implemented.
Main semantic change is that C allows "-1" to be converted into
an unsigned value.  This seemed inconsistent with Chapel philosophy
to me (in particular, Steve points out that converting "3e10" to
an unsigned value would result in "3" rather than a big integer,
as justification for why we won't take a numeric interpretation
of the string and then convert as opposed to simply taking the
string to the correct value.  Updated one test of Steve's that
had previously tested this, but wasn't particularly relying on it
being correct.

In terms of interface, this change changed the *_precise string
conversion routines by now returning two values by pointer argument:
the valid flag as before, and the character that caused the parsing
to choke (in case the client wants to print something useful about
that character -- our cast routines now do print this out as
part of their error messages).

Next improvement that should be made here is to add filename/line#
information to the errors that are generated to help the user
connect the problem to the source code.

I also expect that there are other conversions that we may want
to permit but which are not permitted today (whitespace being a
key case; I'm now realizing that I probably disabled "+3" for
uints in addition to "-3").  I didn't dive into either of these
yet.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11725 3a8e244f-b0f2-452b-bcba-4c88e055c3ca