Added support for promotion over generic functions such that the promotion happens and the generic function is instantiated with respect to the scalar type.  This works because we first attempt to instantiate the generic function over the whole type and then only attempt to instantiate the generic function over the scalar type if the first attempt failed.  Removed test_promote_generic.future.
Fixed a bug in which canDispatch would not flag an argument dispatch
as requiring promotion if the argument was first coerced.

Fixed a bug in which references would break the determination of
subtypes as figured by using : in where clauses.

Added call disambiguation support for functions that require promotion
and those that do not.  Promotion is used as a tie breaker if neither
function is a better match according to the consideration of all of
its arguments.  This tie breaker is used before scope disambiguation.
Updated the spec.  Removed arrayVsPromotion.future and
rangeVsPromotion.future.

Removed support for scalar op tuple and tuple op scalar intrinsics.
All that is left is tuple op tuple intrinsics.  Updated the spec.
This change comes about after much conversation with Brad and David.
These excess functions create problems for promotion and arrays of
tuples.  Let t be a tuple and A be an array of tuples.  The call t+A
would match to a tuple op scalar intrinsic since that doesn't require
promotion.  I believe this simplification is a good step forward.  The
user or domain-specific library writer can do what they want, but for
the general language, this seems to lead to trouble.  Updated several
types/tuple/diten tests.  Updated the spec.  Made a new copy of
jacobi5 called jacobi5a that is a major improvement (removing the
nested definition of + over two 2*int values).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11766 3a8e244f-b0f2-452b-bcba-4c88e055c3ca