Moved man page sources from doc/man to man/ and made the top-level Makefile build the man pages as part of the default build rule for developers. The intention here is to (a) make the man page more easily usable by developers (by adding $CHPL_HOME/man to our MANPATH), and (b) encourage us to keep it up-to-date.
Moved the txt2man script into the third-party
directory and moved the grabVersion.pl perl
script that determined the compiler version from
the doc/man directory into the util/ directory.
Updated the gen_release script to use this new
structure, but haven't tested yet, as the files
need to be in CVS in order to do so.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11774 3a8e244f-b0f2-452b-bcba-4c88e055c3ca