Added a developer-only flag --test-int-sizes used to verify that the compiler's internal integers are of the appropriate size.  Added a test of this to make sure that it's correct on all of our platforms.
Made a new test directory for this test called
"portability".  The idea behind this directory is
to put tests in it that we might want to use as
a sanity check on a new platform to make sure
things work well there (without running all tests,
if we don't have time to) -- tests of timers and
other benchmarking-sensitive details would also
be appropriate to put here.

Next step is to modify the build procedure to
automatically define the compiler's int sizes
and format flags appropriately for a given
platform.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11790 3a8e244f-b0f2-452b-bcba-4c88e055c3ca