[Apologies in advance for making this change after David has gone home and unable to update his cron jobs.  I started the change, realized that I either needed to commit it or have my cron jobs fail, and also figured it would be nice to get this all settled out before the weekend's cron jobs ran.]
Split CHPL_PLATFORM into CHPL_HOST_PLATFORM
and CHPL_TARGET_PLATFORM.  If unset, CHPL_HOST_PLATFORM
defaults to our best guess and CHPL_TARGET_PLATFORM
defaults to CHPL_HOST_PLATFORM.  To cross-compile,
CHPL_HOST_PLATFORM should be set (or inferred) based
on the platform on which the cross-compilation is
occurring, while CHPL_TARGET_PLATFORM needs to be
set to describe the platform being compiled for.

Renamed the util/platform script to util/platform.pl
as a means of making this change fail badly if it's
going to fail, and made this script take an option
--host or --target flag (if neither is used, it
defaults to --host).  Kept the old util/platform
script around for now, but it should be removed as
soon as we get things working correctly.x

Updated the documentation with respect to these
changes.  Also added some information about CHPL_THREADS
and CHPL_COMM.

Updated the sample dotfiles that we maintain.

Updated the Makefiles and testing system


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11793 3a8e244f-b0f2-452b-bcba-4c88e055c3ca