Improved error message reporting for config vars so that in place of the filename:lineno part of the message, we get:
        <command line parsing of 'x'>

Did this by modifying the filename of the expression
to do the configuration variable conversion, and by
modifying the runtime error message reporting system
so that if a filename is given but line number is 0,
it will just print the filename.

Removed the future relating to this for integers.

The other invalid config var tests should fall out
from this, aside from the fact that we don't do any
error checking for those string conversions (I only
implemented integer conversion improvements this
week).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11825 3a8e244f-b0f2-452b-bcba-4c88e055c3ca