More changes to FTree and MRA to abstract more of the indexing and iterating into the s and d fields of the Function class. This version is looking more independent of the specific data structures used in FTree.
Now, the methods in FTree and in Function take tuple arguments
(curNode: 2*int instead of n and l).  And,
const child = s.get_children(curNode) is used to get the children
of curNode.  child(1) is the left child and child(2) is the right
child.  The index arithmetic with n and l to compute the children is
no longer visible in the code.

More changes are required to diff and diffHelper.  (These routines
ran into some function resolution problems when adding the tuple
argument.)  There are also some stray references to n and l in
get_coeffs that have yet to be removed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11828 3a8e244f-b0f2-452b-bcba-4c88e055c3ca