Added tighter runtime string-to-type conversions for booleans, floats, imaginaries, and complexes to complement the work done on integers this week.  The motivation for this was the remaining configVarInvalid* tests that were outstanding, as challenged by Steve. And the implementation was similar enough that doing it while this code was fresh made sense.
This not only makes the checks for these conversions
tighter, but it also (I hypothesize, but have not
checked) allows for a much richer set of variations
on the strings that legally convert to complexes.
An interesting smattering of choices shows up in
stringToComplex.chpl.  In fact, I challenge the team
to come up with strings that don't convert legally
into complexes (trailing whitespace doesn't count --
currently none of our string-to-type conversions
permit trailing whitespace).

Changed Steve's test_string2prims.chpl test and output
to match my beliefs as to what should happen in some
cases that were previously permitted; Steve if you
disagree with my choices (or know them to conflict
with the spec), please discuss with me.

The main challenge to implementing this was that
(prior to this change) we did the string-to-complex
conversions in the Chapel code; but with the tightening
up of the float and imag string-to-* routines, this
module code didn't work anymore.  And moving this code
into the runtime (which is the cleaner place to
implement it due to Chapel's poor support for
sscanf-style operations) wasn't trivial because the
complex types are code generated (and therefore difficult
to refer to in the runtime).

So, added a new pragma, "defined in runtime" which,
when applied to type symbols, causes them not to be
code generated.  Currently, it's only used for the
complex type.  And, added the complex type definitions
for 64-bit and 128-bit complexes to the runtime.  I
think this may currently make the assumption that those
symbols won't get renamed in the compiler -- we may
need to use this pragma in the symbol renaming code as
well, if the user tries to redefine these symbols...

Converted the string-to-complex cast to be more like
a normal string-to-primitive cast in the standard
modules.

Added a bunch of tests of legal and illegal complex
conversions, and illegal imaginary conversions.

Removed the three futures that this test set out to
fix.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11836 3a8e244f-b0f2-452b-bcba-4c88e055c3ca