Updated CMO_array.chpl test to work with updated methods in SingleLocaleArithmeticDomain and SingleLocaleArithmeticArray. This test now compiles and runs correctly.
More testing is needed to ensure that CMODomain and CMOArray
as written in this test truly support slicing and rank-changes.
For now, I wanted to update the test to familiarize myself with the
changes that were made to the arithmetic arrays and domains.
(This test really doesn't do much to test the CMODomain and
CMOArray.  It would be better to put these classes in with the
standard modules as a distribution and test them thoroughly
as an array distribution.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11840 3a8e244f-b0f2-452b-bcba-4c88e055c3ca