(partnered with deitz)
Returned to pairing with Steve and got back to one of
the unanswered questions from our last pairing session:

	- why does the compiler think the function in this test
	  is generic?

Turned out that the front-end made any formal argument with
no specified type into a dtAny, implying that the argument
was generic.  This change tightens up that case to say that
if it has no specified type and no default value given then
it is a dtAny (otherwise, it has a type which should be
inferred later from the default value).

Added a test case which tried to take advantage of the fact
that the function was generic to call it with a floating
point actual when the formal should have been integer only.
Turns out that this test worked correctly (i.e., failed)
before our check-in (for unknown reasons), but with our
change the error message is much better, saving us from
adding an error message future.

At any rate, this is a good change, as it reduces the
number of generic functions that would've been created --
always a good thing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11852 3a8e244f-b0f2-452b-bcba-4c88e055c3ca