(partnered with deitz)
Added a new boolean flag to the ArgSymbol class --
initUsingCopy -- which indicates whether the argument
should be initialized using a copy from its actual,
or whether it should be initialized using a normal
initialization followed by an assignment from its
actual.  Previously, the latter was always done; with
this change we use the direct copy for formal arguments
whose types are not specified by the user, but for
which default values exist from which the type is
taken.

In particular, this helps the case when a formal
argument is intended to be an array type, but its
size is not declared.  Previously, the generated
code would allocate the array with an empty/unknown
domain, then try to loop over that domain to assign
the array, causing a runtime seg fault.  Now, the
array returned by the function is merely copied
into place and everything is peachy keen.

This fixes the future defaultArrViaFn (which shows
this case very cleanly) and gets us one step closer
to solving some of Jim Dinan's array-within-class
errors.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11866 3a8e244f-b0f2-452b-bcba-4c88e055c3ca