(partnered with deitz)
Added similar "initUsingCopy" logic to the building
of a class constructor as in our previous change
in order to support array initializations in classes
where the array member type is inferred (as in
some of Jim Dinan's tests).

This check-in is somewhat unsatisfactory because we
postulated that the transformations to the
constructor's arguments could be left to normalization
rather than specialized here.  And leaving them to
normalization did seem to cause the same transformations
to occur, yet the compiler would break for a dozen
tests in one of two places for reasons that we weren't
able to figure out.  So we decided to check this in
instead since it is working and work on the mystery
on the side.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11874 3a8e244f-b0f2-452b-bcba-4c88e055c3ca