As a result of Steve's "disambiguating overloaded functions via promotion vs. generics" improvement of the other week, I was able to greatly clean up the range math functions in ChapelRange.chpl such that they no longer were explicitly defined for each integral type, but instead could be defined for the "integral" generic type (which was the original motivation for the feature request futures that Steve's check-in resolved).
Unrelated to the above, other than that it uses ranges:

Also, adding some range slicing tests that I've had sitting
in my directory for awhile, waiting for time to decide what
the correct results of them should be.  Since then, Steve
figured out what the results should be, and implemented that
behavior.  I checked that it made sense for these tests,
created their output files, and am checking them in.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11938 3a8e244f-b0f2-452b-bcba-4c88e055c3ca