David correctly pointed out that our error message for unrecognized flags had gotten incrementally worse as a result of my config var error checking change of yesterday (I knew that it made a few cases a bit worse, but had only done targeted testing, so hadn't updated that test, and was too lazy to do a better implementation that would generate a better error message).
Shamed by his mail, I dove into better handling
of error conditions in configuration variables,
and also tried to improve the error messages
themselves a bit.  All-in-all, I think this puts
our config variable handling in better shape
than it has ever been.

(Note that part of the reason that both the old
and yesterday's error messages weren't great was
because of the fact that they had originally
been written assuming that configuration variables
were only set using -s; then at some point we
added the ability to set them using --, which made
it much vaguer as to whether a bad flag was being
used (say, a mistyping of "--help") or whether the
user was setting a config var inappropriately.

Updated .good files and added several new tests to
make sure that all the cases that this checkin is
meant to cover are working.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11942 3a8e244f-b0f2-452b-bcba-4c88e055c3ca