Moved range-to-domain assignment function to ChapelArray from ChapelRange.  This function allows a range to be assigned to a 1D domain.
Refactored range code slightly (it's a start).

Reimplemented the translate,interior,exterior,expand functions to work
on the range's integral type rather than on int.  Removed _ in front
of these methods figuring they may as well be exposed to the user
especially now that range is a type in Chapel.

Added inline pragmas to the array access functions for the array
wrapper and SingleLocaleArithmetic.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11964 3a8e244f-b0f2-452b-bcba-4c88e055c3ca