REGRESSION NOTE:
 [Error matching program output for memory/shannon/jacobi]
 [Error matching program output for memory/shannon/printMemTable]
 [Error matching program output for memory/shannon/printMemTable2]

 These tests are now failing because we may create extra arrays during
 the process of creating one.  See below.

Improved support for arrays and domains in classes greatly.  This
change may impact the semantics of constructors in several ways but
constructors are so poorly defined, documented, and understood that
there is no reason this should be a concern.

Removed 13 futures related to array and domain fields (updating two
goods with the correct output).

Made it so that value information specified in a record type
specification does not get propagated to the constructed record.  See

  types/type_variables/jplevyak/construct-1.good

for an example of this change.

Removed support in the compiler for specifying a type of a field using
a generic type and updated the spec.  We used to allow something like:

  class C { type t; ... }
  class D { var c: C; ... }

Since C is generic, the field c in D can no longer be specified to be
of type C.  Futurized

  classes/bradc/paramInClass/weirdParamInit4.future

which is the only test that attempts such.  The error message is
decent but refers to a standard module for an unknown reason; it is a
future awaiting a better error message.

Added a _set_field function that is used to copy, as necessary, a
variable before setting the member of a class or record in the
constructor.  This is specially defined for sync variables but the
semantic issues related to sync fields and the constructor needs
further study.

Removed the ArrayTypeInfo record which used to serve as a runtime
array type (but only in a few places).  This impacts performance in
that we now generally create extra arrays in their entirety when
building them.  See regression note above.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12036 3a8e244f-b0f2-452b-bcba-4c88e055c3ca