Updated the list of reserved words in the spec, getting it more in sync with the chapel.lex file. A few things are given as reserved words, though seem like they shouldn't need to be -- nil, and compilerError.  I left them on for now.  A few other things are in the lexer, but not in the list (_extern, __primitive) and I left them off for now, figuring the former was an undocumented feature and the latter was for internal use only.
I also reformatted this table in the source and
in the document so that the columns read down
rather than across.  I found this easier to read
as a user, and it should also be easier to maintain
(as I've implemented it with a tabular environment).

I used a sneaky trick of an invisible column in
the table (with white characters) to make the
columns the same size.  I still can't believe that
Latex's tabular environment doesn't support a /kill
line like its tabbing environment does.  Would be
ideal for cases like this.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12052 3a8e244f-b0f2-452b-bcba-4c88e055c3ca