Add uint(64) op signed int operators to the ChapelBase module. There are six versions for each of the operators + - * / ** % == != > < >= <=. Operating over uint(64) and int causes a compilerError except when the int is a parameter and is positive.
Also, remove the check that was built into the compiler that disallows calling
an operator defined over a uint(64) and signed int. Now, the default operators
can be overridden for uint(64)/int combos in user code, for example to define
them more like C, if for some reason they wanted that. :)

I generated the operators with util/gen_int_uint64_operators.chpl, which uses
util/operator_defs as an input file. For future updates to these operators,
it would be good to update the generator instead of the functions directly.
This is mentioned in a comment in ChapelBase.chpl, though maybe not prominently
enough.

Add tests of some operators over positive and negative param values and a test
that overrides operators at module scope.

Update a few .goods for trivial changes in error message wording.

There is one regression: studies/hpcc/RA/bradc/ra-condexpr.chpl
This was expecting the check I removed in the compiler to catch uint(64) ^ int.
Instead, it is now an unresolved call. I plan to define the 12 bitwise
operators Brad has been waiting for next, which will take care of this.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12056 3a8e244f-b0f2-452b-bcba-4c88e055c3ca