Improving the string->complex conversions to yield a better error message for a case that David checked in.  Previously, I'd felt hog-tied by the note in the sscanf() man page indicating that one shouldn't rely on the effect of %n on sscanf's return value.  Since then, I realized that I could query whether or not an sscanf implementation counted %n matches or not, and use that information to correctly gauge the number of matches.  This allowed for better parsing and error reporting.
Also removed a future of Steve's, which I browbeat him into
accepting as working properly.

Next step for all of this string->type conversion stuff is
to make these conversions more open to leading and/or
trailing whitespace, and to implement incremental versions
of them that grab out the specified type and then return
an offset where they left off, so that they can be chained
together to implement file I/O, e.g.  (On our path to supporting
Readers and ReadThis() methods and the like...)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12075 3a8e244f-b0f2-452b-bcba-4c88e055c3ca