Reworked how we can handle var functions that return array and domain wrappers.  Removed Symbol::canVarReturn.  Changed "valid lvalue" pragma to "valid var" pragma.
Added code to the cullOverReferences pass to change functions that
return references to array and domain wrappers into functions that
return array and domain wrappers (not references).  This is necessary
because certain functions in the modules that have a "valid var"
pragma build up a new array or domain wrapper and then return it by
reference.  The "valid var" pragma is used to tell the compiler not to
check the correctness of the returned expression to make sure it can
be a legal reference.  The handling introduced in this checkin
actually removes the reference.

Added four test cases involving the implicit setter argument on var
functions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12119 3a8e244f-b0f2-452b-bcba-4c88e055c3ca