Given that we currently support var and param functions and are on the cusp of supporting type functions, it seemed as though our current functions might be considered const functions by default (and in fact, in most cases, they already are).  At first blush, this makes a lot of sense to Steve and me.
Under this assumption, adding 2 tests that
currently work "correctly" -- returning
ints and classes from functions -- and 2
that work "incorrectly" -- returning records
(which isn't really interesting since
you can't capture the record once you've
modified it, so might as well make it const)
and arrays (which arguably are the only
interesting case, but which I'd like
to make const for consistency and simplicity.
These are currently labeled as semantic
futures.

Also adding a case requesting that the
const keyword be allowed to be explicitly
specified for orthogonality, regardless
of the above proposal.  If the above
proposal is agreed upon, the implementation
of this is easy -- simply do what we normally
do in the parser for functions with no
param/type/var keyword.

See the .future files for more details.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12125 3a8e244f-b0f2-452b-bcba-4c88e055c3ca