Here are two tests of arrays in classes/records that I was writing yesterday in response to Dan Grossman's email, but couldn't get to compile due to two different internal errors.  I'm not sure what makes these tests different than other array-in-class stuff we've done.
These tests also raise a question about whether or not
an element of an array field should be modified when
a record is passed by const (blank) intent.  If we
decide this should be illegal, we should split these
tests into two versions so that the other legal runtime
code can be tested.

A few more notes in the futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12149 3a8e244f-b0f2-452b-bcba-4c88e055c3ca