Here's a future regarding where arrays are copied and where they are referenced with tuples.  It's a variation of what Dan Grossman sent out yesterday that exercises a few other cases.  It ended up messing with my mind, so I'm filing it as a semantic future.
From the .future:

It's no longer clear to me what the semantics of arrays in
tuples should be, both at construction time and when passing
them to a function.  I think Steve's got a good intuition
based on recent 590o email discussions, but wanted to capture
this test and question while I had code showing it off.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12151 3a8e244f-b0f2-452b-bcba-4c88e055c3ca