Paired with David Iten
Made it so that passing a value to a function when a type is expected
is no longer legal (with the exception of constructors which look like
types).

Added Symbol::isConstructor to mark temporaries as capturing
constructors and thereby allowing them to be passed to functions where
a type is expected.  Added disambiguation code to favor values over
types when both are viable.  To turn it into a type, use ".type" on
the constructor call.  Added a pragma "constructor type" to wrappers
of constructors so that wrappers do not interfere with the setting of
the isConstructor field.

Improved support for type propagation through temporary symbols.

Fixed some code that is generated during parsing that relied on
passing values to types.

Fixed several bugs that emerged in test programs as a result of this
change, including Madness and several NAS benchmarks.  Nothing very
deep here, however.

Added inline pragmas to bitRotLeft and bitRotRight to support a faster
bit matrix multiply implementation that David and I also wrote up
today.

Changed build_domain_type and build_index_type into type functions so
that domains and arrays can be built with element types that are
domain or index types.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12171 3a8e244f-b0f2-452b-bcba-4c88e055c3ca