Added an overloaded definition of * to the standard modules that is a type function returning a homogeneous tuple.  It takes as its arguments an integral parameter and a type.  This allows homogeneous tuple types to be used anywhere in the code, not just as a syntactic sugar at the top-level.  Updated the spec.
I attempted to remove the special-casing rule in the parser that
handles top-level homogeneous tuple types in favor of resolving this
function but ran into an ambiguity with range arithmetic.  Go figure.
The issue is that the specification "rank*range(idxType)" results in
an ambiguity between range arithmetic and this new type function.  The
first argument is a better match to the type function since it is a
parameter.  The second argument is a better match to the range
arithmetic function since this "constructor" is preferred as a value
over a type.

The ambiguity can be resolved if the user writes the more verbose
"rank*range(idxType).type" but this doesn't work for formal arguments
or class fields since the type expression is not normalized (yet).
Alternatively, we can define range(...) to only be a type, not a
value.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12179 3a8e244f-b0f2-452b-bcba-4c88e055c3ca