Fixed a bug that Aniruddha ran into in iterating over associative arrays after elements had been deleted from them -- the problem (and solution) mirror a similar bug that he found awhile back when iterating over associative domains.
This exposed a previously undiscovered bug in Madness
in which a computation (integral?) over a constant 1.0
function had resulted in something lower than 1.0,
attributed to roundoff errors.  Hand-inspection shows
that the iteration previously used an empty slot due
to this bug -- the fix now results in the 1.0 result.
(When Jim gets back from India, he should take a look
at this and double-check me -- unless Aniruddha or
M-B can/want to).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12193 3a8e244f-b0f2-452b-bcba-4c88e055c3ca