Added a "runtime array type" called _ArrayTypeInfo that stores the domain value and passes it to functions that are instantiated over an array.  This also gets rid of the extra array allocations made when creating an array (removing two of three linux64 regressions).  The remaining one has to do with an array field and requires some extra work.
Added a build_array primitive used as a place-holder for converting a
runtime array type into an array.

Added two maps to function resolution.  The buildArrayMap maps
build_array primitive calls to instantiations of the "buildArray"
method which are inserted after function resolution when the runtime
array type is inserted.  The wrapDomainMap maps domain classes (not
the wrapper record) to instantiations of the _wrapDomain function
which builds a domain wrapper around a domain class.

Added "rank" fields to associative and enumerated domains so that
wrappers can easily be built around all domain classes by assuming
they have a field named rank.

The function buildArrayTypeInfo builds a record that stores a runtime
array type.  The runtime array type contains the domain and the
element type, which in turn may be another runtime array type (as in
the case of an array of arrays).  This checkin goes a long way to
getting arrays of arrays working, though it is not quite there yet.  I
have not diagnosed what is wrong, but I believe the problem has to do
with storing a runtime array type in a class.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12203 3a8e244f-b0f2-452b-bcba-4c88e055c3ca