Most of this weekend's regressions on GASNet and Valgrind were due to the runtime array type that Steve introduced on Friday.  It contains a domain field and an "element type" field, which is only used at runtime when the element type is another array and not a scalar type (I believe, and this change seems to verify that belief).  The GASNet and Valgrind problems that I investigated were cases of static or runtime complaints about that field being uninitialized. This change eliminates the field when the element type is not an array.  Steve will definitely want to check this when he returns, but my hope is to clean up the regressions until that point.
At first, I took the somewhat hacky approach of assocating a
new pragma with the element type field (named "elt") causing
it to be squelched at code generation time if the field was
not an array.  This worked for all tests but two -- one a
--baseline test against jacobi, for which optimizations didn't
fold in an initialize call for an ArrayTypeInfo call; a
second in which GC had already inspected the field, resulting
in code referencing it being generated.  Lesson learned: if
you don't want a field in the generated code, eliminate it as
quickly as possible.

To this end, I took a somewhat different hacky approach and
made the code that I found in function resolution which
processed the "elt" field and checked whether it was an
array or not go one step further: avoiding creating code
initializing that field if it was not a record, and removing
the field.  This resolves the GASNet warnings that I spot-
checked and keeps the normal linux64 run regression-free.
We'll see what effect it has on the larger GASNet/Valgrind
runs tonight.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12213 3a8e244f-b0f2-452b-bcba-4c88e055c3ca